refactor client folder layout (#1949)
* refactor client folder layout

* split build envs

* fixed import

* excluded folder
