+++ conans/__init__.py
@@ -3,14 +3,14 @@
 from conans.model.conan_file import ConanFile
 from conans.model.options import Options
 from conans.model.settings import Settings
-from conans.client.cmake import CMake
-from conans.client.meson import Meson
-from conans.client.gcc import GCC
-from conans.client.configure_environment import ConfigureEnvironment
-from conans.client.configure_build_environment import (AutoToolsBuildEnvironment, VisualStudioBuildEnvironment)
+from conans.client.build.cmake import CMake
+from conans.client.build.meson import Meson
+from conans.client.build.gcc import GCC
+from conans.client.build.configure_environment import ConfigureEnvironment
+from conans.client.build.autotools_environment import AutoToolsBuildEnvironment
+from conans.client.build.visual_environment import VisualStudioBuildEnvironment
 from conans.client.run_environment import RunEnvironment
 from conans.util.files import load
-import os
 
 # complex_search: With ORs and not filtering by not restricted settings
 COMPLEX_SEARCH_CAPABILITY = "complex_search"
