+++ conans/client/conf/detect.py
