+++ conans/test/integration/conf_default_settings_test.py
@@ -5,7 +5,7 @@ from collections import OrderedDict
 
 from conans import tools
 from conans.client.client_cache import ClientCache
-from conans.client.detect import detect_defaults_settings
+from conans.client.conf.detect import detect_defaults_settings
 from conans.paths import CONANFILE_TXT
 from conans.test.utils.test_files import temp_folder
 from conans.test.utils.tools import TestClient
